Vectorise W assignment in case 'DiDe_with_rise_time'
Vectorise assignment and test equality of results.
One assertion fails because vectorised sumTerm uses 'omitnan' option
in sum function -- this isn't the same as the original code which
discards summing the whole column if it contains a NaN.

Next commit will correct this - just splitting up commits to show
that I've considered it.