Confirm icsig is equal to original calculation
Due to the original conditionals calculating sumTerm,
there wasn't an easy way to assert sumTerm and sumTerm_vec are
the same. Sometime the assertion would fail when sumTerm_vec was an
empty 8*0 array, and sumTerm was 0.

However this new assert confirms that icsig (which is calculated using
sumTerm) is the same as the original method.